+++ sdk/openai/openai/samples-dev/toolCall.ts
@@ -41,7 +41,7 @@ export async function main(): Promise<void> {
   const scope = "https://cognitiveservices.azure.com/.default";
   const azureADTokenProvider = getBearerTokenProvider(new DefaultAzureCredential(), scope);
   const deployment = "gpt-4-turbo";
-  const apiVersion = "2025-03-01-preview";
+  const apiVersion = "2025-04-01-preview";
   const client = new AzureOpenAI({ azureADTokenProvider, deployment, apiVersion });
   const result = await client.chat.completions.create({
     messages: [{ role: "user", content: "What's the weather like in Boston?" }],